requirements.txt
bertopic==0.16.0
black==23.12.1
ipykernel
ipython
ipywidgets==8.1.1
jupyter==1.0.0
matplotlib==3.8.2
matplotlib-inline 
nltk==3.8.1
openai==1.6.1
pandas==2.1.4
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
tiktoken==0.5.2
tokenizers==0.15.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
transformers==4.36.2
python-dotenv
chromadb
datasets
azure-ai-generative[evaluate]
mlflow==2.11.3
keyring==24.3.1
keyrings.alt